feat: add departments' DTO

+++ src/main/java/com/example/UserDept/services/DepartmentService.java
@@ -1,20 +1,19 @@
 package com.example.UserDept.services;
 
-import java.util.List;
-import java.util.Optional;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.UserDept.entities.Department;
 import com.example.UserDept.repositories.DepartmentRepository;
-import com.example.UserDept.exceptions.DatabaseException;
 import com.example.UserDept.exceptions.ResourceNotFoundException;
 
 
-
+@Slf4j
 @Service
 public class DepartmentService {
 
@@ -22,28 +21,26 @@ public class DepartmentService {
 	private DepartmentRepository repository;
 	
 	@Transactional(readOnly = true)
-	public List<Department> findAll(){
-		return repository.findAll();
+	public Page<Department> findAll(Pageable pageable){
+		return repository.findAll(pageable);
 	}
+
 	@Transactional(readOnly = true)
 	public Department findById(Long id){
-		Optional<Department> obj = repository.findById(id);
-		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
+		return repository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	@Transactional
 	public Department insert(Department obj) {
+		log.info("Criando departamento: {}", obj.getName());
 		return repository.save(obj);
 	}
 	
 	@Transactional
 	public void delete(Long id) {
-		findById(id);
-		try {
-			repository.deleteById(id);
-		}catch (DataIntegrityViolationException e) {
-			throw new DatabaseException(e.getMessage());
+		if(findById(id) == null){
+			throw new ResourceNotFoundException(id);
 		}
+		repository.deleteById(id);
 	}
-	
 }
